Transactional Report Ordering, Ranking and Cleanup
Ordering and Ranking details are stored in cell_agg_render_ref column of table
named report_dyn_trans_agg_def.
Pandas Data Frame of the qualified data is created and sorting and ordering is
done on that Data Frame and the result is stored in table report_dyn_trans_summary.